+++ CHANGES.md
@@ -6,12 +6,15 @@ Features:
 
 Improvements:
  - UI for pending edits (#193)
+ - UX image preview screen transition (#393)
 
 Other changes:
  -
 
 Bugfix:
  - Edited message: link confusion when (edited) appears in body (#398)
+ - Close detail room screen when the room is left with another client (#256)
+ - Clear notification for a room left on another client
 
 Translations:
  -